+++ stdnet/utils/structures.py
@@ -1,6 +1,6 @@
 import sys
 from collections import *
 
-if sys.version_info < (2,7):
+if sys.version_info < (2,7):    # pragma nocover
     from .fallbacks._collections import *