+++ tests/regression/fields/id.py
@@ -3,9 +3,9 @@ from uuid import uuid4
 import time
 
 import stdnet
-from stdnet import test
+from stdnet import test, FieldError
 
-from examples.models import Task, WordBook
+from examples.models import Task, WordBook, SimpleModel
 
 
 def genid():
@@ -68,6 +68,44 @@ Use the manager for convenience.'''
         self.assertRaises(Exception, t.save)
 
 
+class TestAutoId(test.TestCase):
+    model = SimpleModel
+    
+    def setUp(self):
+        self.register()
+        
+    def testMeta(self):
+        pk = self.model._meta.pk
+        self.assertEqual(pk.name,'id')
+        self.assertEqual(pk.type,'auto')
+        self.assertEqual(pk.internal_type,'numeric')
+        self.assertEqual(pk.python_type,int)
+        self.assertEqual(str(pk), 'examples.simplemodel.id')
+        self.assertRaises(FieldError, pk.register_with_model,
+                          'bla', SimpleModel)
+        
+    def testCreateWithValue(self):
+        # create an instance with an id
+        m1 = SimpleModel(id = 675, code = 'bla').save()
+        self.assertEqual(m1.id, 675)
+        self.assertEqual(m1.code, 'bla')
+        m2 = SimpleModel(code = 'foo').save()
+        self.assertEqual(m2.id, 676)
+        self.assertEqual(m2.code, 'foo')
+        qs = SimpleModel.objects.query()
+        self.assertEqual(qs.count(),2)
+        self.assertEqual(set(qs),set((m1,m2)))
+    
+    def testCreateWithValue2(self):
+        m1 = SimpleModel(code = 'bla').save()
+        m2 = SimpleModel(id = 575, code = 'foo').save()
+        self.assertEqual(m1.id, 1)
+        self.assertEqual(m2.id, 575)
+        qs = SimpleModel.objects.query()
+        self.assertEqual(qs.count(),2)
+        self.assertEqual(set(qs),set((m1,m2)))
+    
+    
 class CompositeId(test.TestCase):
     model = WordBook