+++ CHANGELOG.rst
@@ -43,7 +43,7 @@ Ver. 0.7.0 - Development
   using python 2.6.
 * Moved the contrib module to :mod:`stdnet.apps`.
 * Added :mod:`stdnet.utils.dates`.
-* **548 regression tests** with **88%** coverage.
+* **553 regression tests** with **88%** coverage.
 
 .. _vers06: